+++ src/components/button-group/index.ts
@@ -1,0 +1,24 @@
+import { createElement, ReactNode } from 'react';
+
+import { IStyle, useClassnames } from './../../hooks/use-classnames';
+import style from './index.module.pcss';
+
+export interface IProps {
+    className?: IStyle | string,
+    direction?: 'row' | 'column',
+    presetStyle?: 'desktop' | 'mobile',
+    children?: ReactNode,
+    tagName?: string
+}
+
+export const ButtonGroup = ({ direction = 'column', tagName = 'div', presetStyle = 'desktop', ...props }: IProps) => {
+    const cn = useClassnames<typeof style>(style, props.className);
+
+    return createElement(tagName || 'div', {
+        children : props.children,
+        className: cn('button-group', {
+            [`button-group_${direction}`]  : direction,
+            [`button-group_${presetStyle}`]: presetStyle
+        })
+    });
+};